Functional Resizing of Wave in '/page.js', Implemented Functional Particle System (Will Be Repolished), Fixed Front Page Design and Button Looks, Implemented 'Under Construction' Page in '/pages/Misc/UnderConstruction'
Performed a lot of front page polishing, updating and improving on looks.

Fixed the resizing issue for the front page waves, so they should show correctly no matter the resolution of the screen.

Finally implemented a working version of the particle system for the front page, as well as a particle system.